--- a/StorScore.ver
+++ b/StorScore.ver
@@ -2,4 +2,4 @@ The results in this directory were generated from
 the version of storscore published on the date listed 
 below.  If there is no date, the results were generated
 from an unpublished & untracked version of code.
-Tue 01/24/2017  9:00:53.83 
+Tue 02/21/2017 16:48:27.05