--- a/lib/DiskSpdRunner.pm
+++ b/lib/DiskSpdRunner.pm
@@ -66,10 +66,7 @@ if( hyperthreading_enabled() )
 
     $affinity_string = join ',', ( @physical_cores, @logical_cores ); 
 }
-else
-{
-    # Do nothing, DiskSpd uses simple round-robin affinity by default
-}
+# Else do nothing, DiskSpd uses simple round-robin affinity by default
 
 sub run($$)
 {
@@ -78,18 +75,130 @@ sub run($$)
     my $test_ref = shift;
     my $run_type = shift;
 
-    my $write_percentage = $test_ref->{'write_percentage'};
-    my $access_pattern   = $test_ref->{'access_pattern'};
-    my $block_size       = $test_ref->{'block_size'};
-    my $queue_depth      = $test_ref->{'queue_depth'};
-    my $run_time         = $test_ref->{'run_time'};
-    my $compressibility  = $test_ref->{'compressibility'};
-  
+    my $run_time = $test_ref->{'run_time'};
+
     if( $run_type eq 'warmup' )
     {
         $run_time = $test_ref->{'warmup_time'};
     }
+    elsif( $run_type eq 'precondition' )
+    {
+        $run_time = '3600' 
+            unless ( $pretend or $self->cmd_line->demo_mode );
+    }
    
+    my $cmd;
+
+    if( $test_ref->{'target_count'} > 1 )
+    {
+        $cmd = $self->get_xml_cmdline($test_ref, $run_time);
+    }
+    else
+    {
+        $cmd = $self->get_cmdline($test_ref, $run_time);
+    }
+
+    my $out_file = $self->output_dir .
+        "\\$run_type-$test_ref->{'description'}.txt";
+    
+    open( my $OUT, ">$out_file" )
+        or die "could not open $out_file: $!";
+   
+    # Save the command line as line 1 of the file
+    print $OUT "$cmd\n";
+
+    my ( $errorlevel, $stdout, $stderr ) = execute_task( $cmd );
+   
+    print $OUT "$stdout\n";
+    print $OUT "$stderr\n";
+
+    close( $OUT );
+
+    print STDERR "\n\tDiskSpd returned non-zero errorlevel" if $errorlevel;
+
+}
+
+sub get_xml_cmdline($$$)
+{
+    my $self = shift;
+    my $test_ref = shift;
+    my $run_time = shift;
+
+    my $compressibility  = $test_ref->{'compressibility'};
+    my $profile_template = $recipes_dir .
+        "\\xmlProfiles\\$test_ref->{'xml_profile'}";
+
+    my $parser = XML::LibXML->new();
+    my $doc = $parser->parse_file( $profile_template );
+
+    for my $timespan ( $doc->findnodes( '//TimeSpan' ) )
+    {
+        my $duration_node = XML::LibXML::Element->new( 'Duration' );
+        $duration_node->appendTextNode( "$run_time" );
+        $timespan->addChild( $duration_node );
+    }
+
+    # Use default unless compressibility is specified by the test
+    my $entropy_file = 
+        Compressibility::get_filename( 
+            $compressibility // $self->cmd_line->compressibility
+        );
+
+    my $entropy_file_size = int Compressibility::FILE_SIZE;
+
+    for my $data_source ( $doc->findnodes( '//RandomDataSource' ) )
+    {
+        my $file_node = XML::LibXML::Element->new('FilePath');
+        $file_node->appendTextNode( "$entropy_file" );
+        $data_source->addChild( $file_node );
+
+        my $file_size_node = XML::LibXML::Element->new( 'SizeInBytes' );
+        $file_size_node->appendTextNode( "$entropy_file_size" );
+        $data_source->addChild( $file_size_node );
+    }
+
+    my $profile_target_nodes = $doc->findnodes( '//Target' );
+    my $profile_target_count = $profile_target_nodes->size();
+    my $recipe_target_count  = $test_ref->{'target_count'};
+
+    my $msg = "Number of targets in recipe ($recipe_target_count)" .
+            "does not match profile ($profile_target_count)\n";
+    die $msg unless( $profile_target_count == $recipe_target_count );
+
+    my $file_num = 0;
+    for my $target ( $profile_target_nodes->get_nodelist )
+    {
+        $file_num++;
+        
+        my $path = $self->target->volume . "\\$TEST_FILE_STUB$file_num.dat";
+
+        my $path_node = XML::LibXML::Element->new( 'Path' );
+        $path_node->appendTextNode( $path );
+        $target->addChild( $path_node );
+    }
+
+    my $profile_file = $self->output_dir .
+        "\\profile-$test_ref->{'description'}.xml";
+    
+    $doc->toFile($profile_file);
+
+    my $cmd = "DiskSpd.exe -X$profile_file";
+
+    return $cmd;
+}
+
+sub get_cmdline($$$)
+{
+    my $self = shift;
+    my $test_ref = shift;
+    my $run_time = shift;
+
+    my $write_percentage = $test_ref->{'write_percentage'};
+    my $access_pattern   = $test_ref->{'access_pattern'};
+    my $block_size       = $test_ref->{'block_size'};
+    my $queue_depth      = $test_ref->{'queue_depth'};
+    my $compressibility  = $test_ref->{'compressibility'};
+  
     # Gradually increase the number of threads as queue_depth
     # increases in order to prevent CPU limitation.
     my $num_threads = 1;
@@ -151,23 +260,7 @@ sub run($$)
         $cmd .= $self->target->file_name;
     }
     
-    my $out_file = $self->output_dir .
-        "\\$run_type-$test_ref->{'description'}.txt";
-    
-    open( my $OUT, ">$out_file" )
-        or die "could not open $out_file: $!";
-   
-    # Save the command line as line 1 of the file
-    print $OUT "$cmd\n";
-
-    my ( $errorlevel, $stdout, $stderr ) = execute_task( $cmd );
-   
-    print $OUT "$stdout\n";
-    print $OUT "$stderr\n";
-
-    close( $OUT );
-
-    print STDERR "\n\tDiskSpd returned non-zero errorlevel" if $errorlevel;
+    return $cmd;
 }
 
 no Moose;