The test operator passes the following command line to test disk 2 using a multi-target workload:

$> StoreScore.cmd --target=2 --recipe=multitarget.rcp

As with previous versions of StorScore, the command line defines the device under test, and the recipe file defines the workload.  New with these changes, it also defines the number of files to create and test.  For example:

multitarget.rcp:

test(
    description  => "Multitarget Sample"
    target_count => 2,
    xml_profile  => "sample.xml",
    initialize   => 0,
    warmup_time  => 5,
    run_time     => 60,
);

Unlike single-target tests, multi-target tests define the workload in a separate xml profile template.  StorScore checks to make sure the recipe points to a profile file whenever the target count is greater than 1.

During each test, StorScore wipes the disk, creates a partition and file system, and then divides the disk space evenly among the number of targets given by the recipe.  It injects a few variables into the xml template, such as file names and run time, exports the xml document to the results directory and then points diskspd at this profile.  (Diskspd.exe -Xmodified-sample.xml)

NB: In the future, I would like to generate the xml document from scratch to reduce user error and frustration.  The XML profiles can be tedious to generate, and they can easily get out of synch with the recipe.

These changes also update the parsed data.  Previously, the excel file printed one row for each test:

Disk  | Description   | W Mix  | IO Size | MB/s   | Avg Latency
-------------------------------------------------------------------------------------
PM953 | 4k Rand Read  | 0      | 4k      | 502.45 | 0.092
PM953 | 4k Rand Write | 100    | 4k      | 235.45 | 1.201
PM953 | 8k Rand Read  | 0      | 8k      | 252.45 | 0.181

Now, each row shows either per-target workload and results or the aggregate workload and results:

Disk  | Description   | Target       | W Mix  | IO Size | MB/s   | Avg Latency
------------------------------------------------------------------------------------------------------
PM953 | 4k Rand Read  | Total        | 0      | 4k      | 502.45 | 0.092
PM953 | Example Multi | Total        | 50     |         | 235.45 | 1.201
PM953 | Example Multi | E:\file1.dat | 0      | 4k      | 285.12 | 0.181
PM953 | Example Multi | E:\file2.dat | 100    | 8k      | 102.48 | 1.589

The outlier and score calculation is based only on the measurements listed in the aggregated rows.

To support this output format, I added a level to the parser's main data structure:

Before:

stats =
{
    “FW version” => “10010L00”,
    …
    “Read Mix” => 100,
    “QD” => 1,
    …
    “MB/s Read” => 1175,
    “IOs Read” => 20398,
    …
}

After:
stats =
{
    “FW version” => “10010L00”,
    …
    “Workloads” =>
    {
        “Total” =>
                 {
                     “Read Mix” => 50,
                     “QD” => 129,
                     …
                 },
        “E:\file1.dat” =>
                 {
                     “Read Mix” => 100,
                     “QD” => 1,
                     …
                 },
       “E:\file2.dat” =>
                 {
                     “Read Mix” => 0,
                     “QD” => 128,
                     …
                 }
    },
    “Measurements” =>
    {
       “Total” =>
               {
                   “MB/s Read” => 285.12,
                   “IOs Read” => 10128,
                   …
               },
       “E:\file1.dat” =>
               {
                   “MB/s Read” => 285.12,
                   “IOs Read” => 10128,
                   …
               },
       “E:\file2.dat” =>
               {
                   “MB/s Read” => 0,
                   “IOs Read” => 0,
                   …
               }
    }
}