@marksantaniello 2017-02-22
Does this enable us to test multiple files on a single volume, or multiple volumes?

The terminology is starting to get confusing.  If the command line specifies --target=2 (a single target, disk number 2) how can the recipe then specify multiple targets?

I looked at one of the XML files and it doesn't seem to specify a file name, just a list of unnamed targets.  I guess this works when you want StorScore.cmd to test multiple files on a single drive?

Can you test what happens when you use a recipe like this and target an existing volume or file (--target=D:\ or --target=D:\myfile)?  Both of those use cases are supported today.

Is it possible for the targets specified in XML to conflict with the command line --target?  What if the recipe "target_count" property doesn't match the XML file?  Why not parse the XML and just count the targets, instead of requiring the user to do this themselves?  Seems like an easy way to avoid a whole class of bug.

Today we have some guard rails to prevent a user from accidentally destroying the wrong disk.  Does that all still work with this?