@lauracaulfield 2017-02-24
That's really odd... GitHub lets me edit *your* comment, but doesn't let me reply directly to it.  Hmmm... what embarrassing thing can I make Mark say on the internet... j/k :-)

This is a big feature to add to StorScore, and I'd like to work out better terminology.  But, it's a useful feature for evaluating a new functionality ("Streaming") that's coming to an SSD near you.  I think I need to do a better job of explaining why StorScore should be able to test that feature. 

The top-level goals are the same as always 
1. apply a set of tests ("recipe") to a single drive with one click and a large amount of computer time
2. process that big pile of data and compare how different drives did on the same set of tests (with varying weights)

The technology change that's motivating the change to StorScore is called streaming (there's some information about it here: https://www.usenix.org/system/files/conference/hotstorage14/hotstorage14-paper-kang.pdf).  Basically, the streaming interface allows the host and drive to communicate about which data should be grouped together on the SSD.  This is really helpful in preventing GC when you're, say, writing each of 4 files sequentially and simultaneously.  In a normal SSD, the 4 streams mix together and create a high WAF.  In a streaming drive, the controller places the data from each stream in it's own set of blocks.  Each stream can write at it's own rate, and the host writes will invalidate all the data in a block at roughly the same time, and WAF will be 1.  

Now, enter StorScore.  The goal is to be able to define a streaming test (like the 4-file example above) in the recipe, and be able to parse the results and compare them across drives.  In the parsed data, I would like to be able to see the performance of each stream, the aggregate performance, and the WAF (of the whole drive) to validate that the streams are implemented properly.  


Now, for some specific answers.  The multi-stream/target/file test is still meant to evaluate a single volume or disk -- the volume or disk specified on the command line.  Volume and disk targets are supported for this type of test, but file is not.  When the user targets a file with this type of test, StorScore should give a useful error message.  

The XML "template" file doesn't list file names (and a few other parameters like the test's run time) because StorScore generates those and adds them to the final profile before running the test.  StorScore gets the size of the target given on the command line, and divides this space evenly among the number files given in the test, and creates the files.  This way StorScore can still purge between tests when the target is a disk.  Also, the user doesn't have to know about the valid data length flag or do preconditioning manually.

If the number of files (currently "target_count" in the recipe, but I'm seriously thinking about renaming it) doesn't match the number of targets in the XML file, StorScore errors.  I prefer to generate XML rather than parse it, because I want StorScore defining the workload rather than the manually generated (read: error prone) XML files.  Ultimately, I would like StorScore to generate the entire XML file from scratch.  I'm using the template idea temporarily until we get a better feel for the knobs we want in this type of test.  

Regarding this:
_Today we have some guard rails to prevent a user from accidentally destroying the wrong disk. Does that all still work with this?_

Are you talking about the initial message "are you sure you want to destroy..."? If so, yes.  This is still functional for all types of tests.  


This is a big design change, but from my perspective it fits well within our original goals.  I'm surprised it didn't require more changes as well, but after working with it, I think that's just because it has a structure that matches our broader goals -- even the ones we hadn't defined yet.  I tested with a conventional test (for StorScore changes) and conventional set of data (for parser changes), so I'm as confident in backward-compatibility as I can be without a regressions test.