@marksantaniello 2017-02-24
I guess DiskSpd's XML file uses the word "target", which is adding to the confusion over terminology.  Would things make more sense if we kind of ignored that and used the word "stream" in some places?  So we still test one "target" with StoreScore, but that "target" can potentially have many "streams" ?