@lauracaulfield 2017-02-24
Stream is a fairly overloaded term I try to avoid.  It might be ok here, though, since StorScore isn't using the term for anything else.  Whatever term we use, I think a clean separation from diskspd's terminology will help.  

Some options:
sub-target
stream
file
channel <-- also over loaded

Technically speaking, stream is the SSD-level name.  You may not always have a one-to-one mapping between stream and file.  Even so, I think stream is still my favorite.