@marksantaniello 2017-02-24
So... one more idea I'd like to suggest to you:

If there's some tension between a "nice design" and "getting something working now", you could just fork StorScore and create StreamingStorScore, for the latter, to break the dependency.

The fancier design could specify multiple streams directly in our recipe format, and delegate to the IO generator how to accomplish it.  For DiskSPD or (someday, I hope) FIO, maybe you would have some code to construct an configuration input file (XML, in the case of DiskSPD).  This could live in DiskSpdRunner/SqlioRunner/FioRunner because it's pretty analogous to the code that translates our recipe into the right command line flags.  If the IO generator doesn't support multiple streams natively (like SQLIO) maybe you could simulate it by launching multiple child processes (similar to the idea of launching multiple precondition.exes).

By forking, you get something out today that vendors can use, etc., but avoid taking on a lot of tech debt that will get in the way later.